OIDC API Authorization

+++ OrdersFunction.cs
@@ -10,35 +10,57 @@ using System.IO;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using System;
+using OidcApiAuthorization.Abstractions;
+using OidcApiAuthorization.Models;
 
 namespace WhereIsMyOrderAPI
 {
-    public class Orders
+    public class OrdersFunction
     {
         IOrderRepository OrderRepository { get; set; }
 
-        public Orders(IOrderRepository orderRepository)
+        private readonly IApiAuthorization _apiAuthorization;
+
+        public OrdersFunction(IOrderRepository orderRepository,
+            IApiAuthorization apiAuthorization)
         {
+            _apiAuthorization = apiAuthorization;
             OrderRepository = orderRepository;
         }
 
         [FunctionName("GetOrdersForUser")]
-        public Order[] GetOrdersForUser(
+        public async Task<ActionResult<Order[]>> GetOrdersForUser(
             [HttpTrigger(AuthorizationLevel.Anonymous, "GET", Route = "GetOrdersForUser")] HttpRequest req,
             ILogger log)
         {
-            log.LogInformation("GetOrdersForUser processed a request.");
+            log.LogWarning("GetOrdersForUser function processed a request.");
+            ApiAuthorizationResult authorizationResult = await _apiAuthorization.AuthorizeAsync(req.Headers);
+            if (authorizationResult.Failed)
+            {
+                log.LogError(authorizationResult.FailureReason);
+                return new UnauthorizedResult();
+            }
+
+            log.LogWarning($"HTTP trigger function UpdateOrder request is authorized.");
             string userId = req.Query["userId"];
             var results = OrderRepository.GetAll().Where(x => x.UserId == userId).ToArray();
             return results;
         }
 
         [FunctionName("CreateOrder")]
-        public async Task<Order> CreateOrder(
+        public async Task<ActionResult<Order>> CreateOrder(
             [HttpTrigger(AuthorizationLevel.Anonymous, "POST", Route = "CreateOrder")] HttpRequest req,
             ILogger log)
         {
-            log.LogInformation("CreateOrder function processed a request.");
+            log.LogWarning("CreateOrder function processed a request.");
+            ApiAuthorizationResult authorizationResult = await _apiAuthorization.AuthorizeAsync(req.Headers);
+            if (authorizationResult.Failed)
+            {
+                log.LogError(authorizationResult.FailureReason);
+                return new UnauthorizedResult();
+            }
+
+            log.LogWarning($"HTTP trigger function UpdateOrder request is authorized.");
             string requestBody = await new StreamReader(req.Body).ReadToEndAsync();
             dynamic data = JsonConvert.DeserializeObject<OrderRoot>(requestBody);
             var newOrder = data.newOrder;
@@ -48,11 +70,19 @@ namespace WhereIsMyOrderAPI
         }
 
         [FunctionName("UpdateOrder")]
-        public async Task<Order> UpdateOrder(
+        public async Task<ActionResult<Order>> UpdateOrder(
             [HttpTrigger(AuthorizationLevel.Anonymous, "PUT", Route = "UpdateOrder")] HttpRequest req,
             ILogger log)
         {
-            log.LogInformation("UpdateOrder function processed a request.");
+            log.LogWarning("UpdateOrder function processed a request.");
+            ApiAuthorizationResult authorizationResult = await _apiAuthorization.AuthorizeAsync(req.Headers);
+            if (authorizationResult.Failed)
+            {
+                log.LogError(authorizationResult.FailureReason);
+                return new UnauthorizedResult();
+            }
+
+            log.LogWarning($"HTTP trigger function UpdateOrder request is authorized.");
             string requestBody = await new StreamReader(req.Body).ReadToEndAsync();
             dynamic data = JsonConvert.DeserializeObject<OrderRoot>(requestBody);
             var updatedOrder = data.newOrder;
@@ -62,11 +92,19 @@ namespace WhereIsMyOrderAPI
         }
 
         [FunctionName("DeleteOrder")]
-        public ActionResult DeleteOrder(
+        public async Task<ActionResult> DeleteOrder(
             [HttpTrigger(AuthorizationLevel.Anonymous, "DELETE", Route = "DeleteOrder")] HttpRequest req,
             ILogger log)
         {
-            log.LogInformation("DeleteOrder function processed a request.");
+            log.LogWarning("DeleteOrder function processed a request.");
+            ApiAuthorizationResult authorizationResult = await _apiAuthorization.AuthorizeAsync(req.Headers);
+            if (authorizationResult.Failed)
+            {
+                log.LogError(authorizationResult.FailureReason);
+                return new UnauthorizedResult();
+            }
+
+            log.LogWarning($"HTTP trigger function UpdateOrder request is authorized.");
             var orderId = int.Parse(req.Query["orderId"].ToString());
             var order = OrderRepository.GetAll().Where(x => x.Id == orderId).FirstOrDefault();
             OrderRepository.Delete(order);
@@ -75,11 +113,19 @@ namespace WhereIsMyOrderAPI
         }
 
         [FunctionName("DeleteOrders")]
-        public ActionResult DeleteOrders(
+        public async Task<ActionResult> DeleteOrders(
             [HttpTrigger(AuthorizationLevel.Anonymous, "DELETE", Route = "DeleteOrders")] HttpRequest req,
             ILogger log)
         {
-            log.LogInformation("DeleteOrders function processed a request.");
+            log.LogWarning("DeleteOrders function processed a request.");
+            ApiAuthorizationResult authorizationResult = await _apiAuthorization.AuthorizeAsync(req.Headers);
+            if (authorizationResult.Failed)
+            {
+                log.LogError(authorizationResult.FailureReason);
+                return new UnauthorizedResult();
+            }
+
+            log.LogWarning($"HTTP trigger function UpdateOrder request is authorized.");
             var orderIdsString = req.Query["orderIds"].ToString();
             var orderIdsSplitted = orderIdsString.Split(',');
             int[] orderIds = Array.ConvertAll(orderIdsSplitted, s => int.Parse(s));
